debt - have ITimerService for browser and electron
