+++ src/vs/workbench/workbench.desktop.main.ts
@@ -66,10 +66,13 @@ import { IUserDataAutoSyncService } from 'vs/platform/userDataSync/common/userDa
 import { UserDataAutoSyncService } from 'vs/workbench/contrib/userDataSync/electron-browser/userDataAutoSyncService';
 import { ITunnelService } from 'vs/platform/remote/common/tunnel';
 import { TunnelService } from 'vs/platform/remote/node/tunnelService';
+import { ITimerService } from 'vs/workbench/services/timer/browser/timerService';
+import { TimerService } from 'vs/workbench/services/timer/electron-browser/timerService';
 
 registerSingleton(ICredentialsService, KeytarCredentialsService, true);
 registerSingleton(IUserDataAutoSyncService, UserDataAutoSyncService);
 registerSingleton(ITunnelService, TunnelService);
+registerSingleton(ITimerService, TimerService);
 
 //#endregion
